Refactor tests
Preparing to add Greek to Latin test
