+++ tests/conftest.py
@@ -4,15 +4,18 @@ from pathlib import Path
 import getpass
 import json
 import os
+import pprint
 import tempfile
 
 import pymongo
 import pytest
 
-from tesserae.db import TessMongoConnection, Text
+from tesserae.db import TessMongoConnection
+from tesserae.db.entities import Feature, Text
 from tesserae.utils import ingest_text
 from tesserae.utils.delete import obliterate
 from tesserae.utils.multitext import BigramWriter
+from tesserae.utils.search import get_results
 
 
 # Make sure that bigram databases are written out to a temporary location
@@ -193,3 +196,121 @@ def minipop(request, mini_greek_metadata, mini_latin_metadata):
         ingest_text(conn, text)
     yield conn
     obliterate(conn)
+
+
+def _build_relations(results):
+    relations = {}
+    for match in results:
+        target_loc = match['target_tag'].split()[-1]
+        source_loc = match['source_tag'].split()[-1]
+        if target_loc not in relations:
+            relations[target_loc] = {source_loc: match}
+        elif source_loc not in relations[target_loc]:
+            relations[target_loc][source_loc] = match
+    return relations
+
+
+def _load_v3_results(minitext_path, tab_filename):
+    tab_filepath = Path(minitext_path).resolve().parent.joinpath(tab_filename)
+    v3_results = []
+    with open(tab_filepath, 'r', encoding='utf-8') as ifh:
+        for line in ifh:
+            if not line.startswith('#'):
+                break
+        for line in ifh:
+            # ignore headers
+            # headers = line.strip().split('\t')
+            break
+        for line in ifh:
+            data = line.strip().split('\t')
+            v3_results.append({
+                'source_tag': data[3][1:-1],
+                'target_tag': data[1][1:-1],
+                'matched_features': data[5][1:-1].split('; '),
+                'score': float(data[6]),
+                'source_snippet': data[4][1:-1],
+                'target_snippet': data[2][1:-1],
+                'highlight': ''
+            })
+    return v3_results
+
+
+class V3Checker:
+    @staticmethod
+    def load_v3_mini_text_stem_freqs(conn, metadata):
+        db_cursor = conn.connection[Feature.collection].find(
+                {'feature': 'form', 'language': metadata['language']},
+                {'_id': False, 'index': True, 'token': True})
+        token2index = {e['token']: e['index'] for e in db_cursor}
+        # the .freq_score_stem file is named the same as its corresponding
+        # .tess file
+        freqs_path = metadata['path'][:-4] + 'freq_score_stem'
+        freqs = {}
+        with open(freqs_path, 'r', encoding='utf-8') as ifh:
+            for line in ifh:
+                if line.startswith('# count:'):
+                    denom = int(line.split()[-1])
+                    break
+            for line in ifh:
+                line = line.strip()
+                if line:
+                    word, count = line.split()
+                    freqs[token2index[word]] = float(count) / denom
+        return freqs
+
+    @staticmethod
+    def check_search_results(conn, results_id, textpath, tabname):
+        v5_results = get_results(conn, results_id)
+        v5_results.sort(key=lambda x: -x['score'])
+        v3_results = _load_v3_results(textpath, tabname)
+        v3_relations = _build_relations(v3_results)
+        v5_relations = _build_relations(v5_results)
+        score_discrepancies = []
+        match_discrepancies = []
+        in_v5_not_in_v3 = []
+        in_v3_not_in_v5 = []
+        for target_loc in v3_relations:
+            for source_loc in v3_relations[target_loc]:
+                if target_loc not in v5_relations or \
+                        source_loc not in v5_relations[target_loc]:
+                    in_v3_not_in_v5.append(
+                        v3_relations[target_loc][source_loc])
+                    continue
+                v3_match = v3_relations[target_loc][source_loc]
+                v5_match = v5_relations[target_loc][source_loc]
+                v3_score = v3_match['score']
+                v5_score = v5_match['score']
+                if f'{v5_score:.3f}' != f'{v3_score:.3f}':
+                    score_discrepancies.append((
+                        target_loc, source_loc,
+                        v5_score-v3_score))
+                v5_match_features = set(v5_match['matched_features'])
+                v3_match_features = set()
+                for match_f in v3_match['matched_features']:
+                    for f in match_f.split('-'):
+                        v3_match_features.add(f)
+                only_in_v5 = v5_match_features - v3_match_features
+                only_in_v3 = v3_match_features - v5_match_features
+                if only_in_v5 or only_in_v3:
+                    match_discrepancies.append((
+                        target_loc, source_loc, only_in_v5,
+                        only_in_v3))
+        for target_loc in v5_relations:
+            for source_loc in v5_relations[target_loc]:
+                if target_loc not in v3_relations or \
+                        source_loc not in v3_relations[target_loc]:
+                    in_v5_not_in_v3.append(
+                        v5_relations[target_loc][source_loc])
+        pprint.pprint(score_discrepancies)
+        pprint.pprint(match_discrepancies)
+        pprint.pprint(in_v5_not_in_v3)
+        pprint.pprint(in_v3_not_in_v5)
+        assert not score_discrepancies
+        assert not match_discrepancies
+        assert not in_v5_not_in_v3
+        assert not in_v3_not_in_v5
+
+
+@pytest.fixture(scope='session')
+def v3checker():
+    return V3Checker
